GROW-177 refactor: 자산 관리 시트 컴포넌트 리팩토링
