+++ frontend/src/widgets/asset-management-draggable-table/atoms/currentSortingTypeAtom.ts
@@ -1,0 +1,9 @@
+import { atom } from "jotai";
+import {
+  FieldSortingType,
+  FieldSortingTypeValue,
+} from "@/widgets/asset-management-draggable-table/constants/fieldSortingType";
+
+export const currentSortingTypeAtom = atom<FieldSortingTypeValue>(
+  FieldSortingType.DESC,
+);
