+++ frontend/src/widgets/asset-management-draggable-table/atoms/sortingFieldAtom.ts
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const sortingFieldAtom = atom<string | null>(null);
