+++ frontend/src/widgets/asset-management-draggable-table/constants/cellMinimumWidth.ts
@@ -1,0 +1,7 @@
+export const cellMinimumWidth = {
+  종목명: 150,
+  수량: 130,
+  구매일자: 136,
+  증권사: 156,
+  "계좌 종류": 120,
+};
