+++ frontend/src/widgets/asset-management-draggable-table/constants/currencyType.ts
@@ -1,0 +1,6 @@
+export const CurrencyType = {
+  KRW: "KRW",
+  USD: "USD",
+} as const;
+
+export type CurrencyType = (typeof CurrencyType)[keyof typeof CurrencyType];
