+++ frontend/src/widgets/asset-management-draggable-table/constants/essentialFields.ts
@@ -1,0 +1,1 @@
+export const essentialFields = ["종목명", "수량", "구매일자"];
