+++ frontend/src/widgets/asset-management-draggable-table/constants/fieldSortingType.ts
@@ -1,0 +1,7 @@
+export type FieldSortingTypeValue =
+  (typeof FieldSortingType)[keyof typeof FieldSortingType];
+
+export const FieldSortingType = {
+  ASC: "asc",
+  DESC: "desc",
+} as const;
