+++ frontend/src/widgets/asset-management-draggable-table/constants/fieldWidth.ts
@@ -1,0 +1,5 @@
+export const filedWidth = {
+  종목명: 12,
+  계좌종류: 9,
+  수익률: 8,
+};
