+++ frontend/src/widgets/asset-management-draggable-table/constants/numberFields.ts
@@ -1,0 +1,13 @@
+export const numberFields = [
+  "수량",
+  "현재가",
+  "배당금",
+  "고가",
+  "저가",
+  "시가",
+  "수익률",
+  "수익금",
+  "매입금",
+  "매입가",
+  "거래량",
+];
