+++ frontend/src/widgets/asset-management-draggable-table/constants/priceInputFields.ts
@@ -1,0 +1,10 @@
+export const priceInputFields = [
+  "현재가",
+  "배당금",
+  "고가",
+  "저가",
+  "시가",
+  "매입금",
+  "매입가",
+  "수익금",
+];
