+++ frontend/src/widgets/asset-management-draggable-table/utils/fieldIsRequired.ts
@@ -1,0 +1,4 @@
+import { essentialFields } from "@/widgets/asset-management-draggable-table/constants/essentialFields";
+
+export const fieldIsRequired = (field: string) =>
+  essentialFields.includes(field);
