+++ frontend/src/widgets/asset-management-draggable-table/utils/fieldItemFactory.ts
@@ -1,0 +1,7 @@
+import { fieldIsRequired } from "@/widgets/asset-management-draggable-table/utils/fieldIsRequired";
+
+export const fieldIemFactory = (field: string, userFields: string[]) => ({
+  isRequired: fieldIsRequired(field),
+  isChecked: userFields.includes(field) || fieldIsRequired(field),
+  name: field,
+});
