+++ frontend/src/widgets/asset-management-draggable-table/utils/fieldTypeIsNumber.ts
@@ -1,0 +1,5 @@
+import { numberFields } from "@/widgets/asset-management-draggable-table/constants/numberFields";
+
+export const fieldTypeIsNumber = (field: string) => {
+  return numberFields.includes(field);
+};
