+++ frontend/src/widgets/index.ts
@@ -13,6 +13,6 @@ export { default as ExplorePortfolio } from "./explore-portfolio/ui/ExplorePortf
 export { default as RichPortfolio } from "./rich-portfolios/ui/RichPortfolio";
 export { default as Summary } from "./Summary/ui/Summary";
 export { default as HomeGuestAccessGuideButton } from "./HomeGuestAccessGuide/ui/HomeGuestAccessGuideButton";
-export { default as AssetManagementDraggableTable } from "./asset-management-draggable-table/ui/AssetManagementDraggableTable";
+export { default as AssetManagementDraggableTable } from "./asset-management-draggable-table/ui/AssetManagementSheet";
 export { default as Footer } from "./footer/ui/Footer";
 export { default as AssetSheetSummary } from "./asset-management-summary-card/ui/AssetSheetSummary";
